Remove OurBroadnessMeasurer.
The reason it was getting different results than regular
BroadnessMeasurer seems to have been because torch gives different
results on a data point when it's evaluated alone versus as part of a
larger set. Could be related to matrix multiplication algorithms doing
different calculations in the two cases. See discussion at
<https://github.com/LuanAdemi/toumei/pull/15>.

Bump num_iters up to 10000, because that makes results seem fairly
consistent for different seeds.